pyproject.toml
[project]
name = "dsp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.9",
    "aws-cdk-aws-bedrock-agentcore-alpha>=2.232.1a0",
    "aws-cdk-lib>=2.232.1",
    "aws-opentelemetry-distro>=0.12.0",
    "bedrock-agentcore[strands-agents]>=1.1.1",
    "boto3>=1.42.4",
    "pydantic-settings>=2.12.0",
    "strands-agents[a2a]>=1.7.0",
    "strands-agents-tools>=0.1.0",
    "yahoo-dsp-agent-sdk",
]

[tool.uv.sources]
yahoo-dsp-agent-sdk = { path = "agent-sdk", editable = true }

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.2",
    "ruff>=0.14.8",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]
addopts = "-v"
